requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
rasterio>=1.3.0
Pillow>=10.0.0
numpy>=1.24.0
